refactor: migrate api controller file.
